(maint) Get inject-classification-tool.sh to work from bin/
for the case of pushing the tool to vmpooler hosts. Originally the
script was written to push the tool to local vagrant hosts on my laptop,
and it relied on scripts present in a pe-[year].[n] format
directory.

Further clean up is needed, this just gets it working so it can be
called generally from my PATH on the slice workstation.