(maint) Lookup platform from host when it is not set
This script was previously mostly used for vagrant instances were I had
pre-defined env variables about platform/version to source. To get the
helper script injected into a ci pooler host without this, just need to
know which platform I'm targeting, so scrapping via facter for now.

All of this should be replaced with Puppet/Bolt...but this will work for
now.