(maint) Add two scripts for interacting with openstack
One to refresh the token so I don't have to export LDAP credentials into
my environment. The other, oscontrol, is for creating/destroying sets of
os servers. Both rely on fog-openstack.

I would have used vagrant for this, but it relies on exported passwords
credentials, and I wanted to make use of the token.