(maint) Changes for flexibility
This adds changes to be able to set the path to the puppet-enterprise-vanagon folder used for builds. It also changes the rerun-pe-install-with-pe-postgresql10-packages.sh file to find the name of the nfs mounted folder, rather than hard coding to jpartlow-src. The README file is also updated with info on building and installing.