(maint) Break out a generic command runner/executor class for testing
Testing tasks is difficult; by having a generic Executor class as the
base for tasks in this module, and a CommandRunner that can be replaced
with a TestRunner instance, it's easier to test that the scripts
themselves are behaving. Having the core task functionality separated
out into modules means those workflows can be tested separately as well.