(maint) Add plan to build a pglogical extension
The assist_vanagon_build plan was sufficient to copy and install
packages necessary for building a pglogical extension. The problem
was that this needed to happen before vanagon was called, and we had to
prep the host vanagon would be working with. Since a default invocation
of vanagon also checks out a host from vmpooler you get stuck with a
bootstrapping problem.

However, you can tell vanagon to operate directly on a host that you've
prepared ahead of time, so this plan handles preparing a host and then
handing it off to vanagon to build the extension.