(maint) Do not apply_prep before injecting_packages
The apply_prep() function ensures that a puppet-agent is installed on
the node. But since we are preparing a PE tarball for later
installation, we want the node to be pristine, since an existing
puppet-agent would interfere with the installer installing the
puppet-agent version bundled in the PE tarball.

The facts plan returns a base set of os facts similar to the os fact in
facter. The only aspect we were relying on which is present is the
Ubuntu distribution codename, so this commit adds the ubuntu_codename()
function to derive it from full version and makes use of that instead.