(maint) Rework RunShellExecutable state
Keeping in a separate context class now that holds the io, level, muzzle
settings. And storing the context in a Thread local. This makes it
simpler to vary output state for multiple threads. Intention is to pave
the way for splitting Threaded output to discrete IO objects to untangle
the streams.