(maint) Add a module to assist with running in Threads
Based on the Thread code in scripts/test-pe-postgresql.rb, this module
provides a simple mechanism for coordinating execution of a block based
on the cartesian product of a hash of arrays that are the set of unique
combinations to execute in threads.

Expects to be working in a class that has already included
RunShellExecutable.