(maint) Send threaded output to separate files
Adds a flag to run_threaded_product() allowing for the output of each
thread to be sent to a specific temp file for the variant combination
we're executing. This allows something lengthy, like an upgrade workflow
test threaded across all the platforms, or whatever other combination,
to execute in parallel, but have streaming output go to the file rather
than be jumbled in the original $stdout.

A RunShellExecutable::StreamingExecutor is added which uses
Open3.popen2e instead of capture2e. This is still dependent on the rate
at which whatever we are running flushes its output. There may be other
kinks in run_shell.rb or threaded.rb to work out as well.