(maint) Have test_migrations handle multiple versions/layouts
and checkout appropriate number of vms. Previously test_migrations
expected to test upgrading from a single version, and the layout
depended on having prepopulated either 1 or 2 vms per platform.

This patch allows you to test the two 2019.2 layouts, mono or
master/database, singly or together, from multiple initial installation
versions.

To save having to compute the number of vms needed ahead of time, it
calculates that for you and creates the vms as needed, then hands them
out in appropraite chunks to the upgrade_workflow plan.