(maint) Call floaty list just once when checking for live vms
Previously called floaty list --active every time we were testing that a
host might be live. Since we check all hosts within the same function,
this patch calls floaty list once and caches the output for the duration
of the function call, saving time when we are checking many hosts.